refactor: modify build to work with tags
I want to be able to push a tag and have it build, instead of create and publish a release.
Additionally, if somebody forks, I want to ensure that their images are named and pushed to their own context.
Finally, since I only build and push on tags, I don't need to restrict when I push with this workflow.